requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
boto3==1.21.24
botocore==1.24.24
certifi==2021.10.8
charset-normalizer==2.0.12
coverage==6.3.2
Django==4.0.3
django-debug-toolbar==3.2.4
django-elasticsearch-dsl==7.2.2
django-seed==0.3.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
elasticsearch==7.13.4
elasticsearch-dsl==7.4.0
Faker==13.3.2
idna==3.3
jmespath==1.0.0
psycopg2-binary==2.9.3
pydantic==1.9.0
PyJWT==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
toposort==1.7
typing-extensions==4.1.1
urllib3==1.26.9
